OOPSI

+++ src/lib/config/config.cpp
@@ -5,11 +5,11 @@ bool get_LUMIN() {
     return true;
 };
 
-int get_LUMIN_LOW() {
+short int get_LUMIN_LOW() {
     return 255;
 };
 
-int get_LUMIN_HIGH() {
+short int get_LUMIN_HIGH() {
 return 758;
 };
 
@@ -18,11 +18,11 @@ bool get_TEMP_AIR() {
     return true;
 };
 
-int get_MIN_TEMP_AIR() {
+short int get_MIN_TEMP_AIR() {
     return -10;
 };
 
-int get_MAX_TEMP_AIR() {
+short int get_MAX_TEMP_AIR() {
     return 60;
 };
 
@@ -31,11 +31,11 @@ bool get_HYGR() {
     return true;
 };
 
-int get_HYGR_MINT() {
+short int get_HYGR_MINT() {
     return 0;
 };
 
-int get_HYGR_MAXT() {
+short int get_HYGR_MAXT() {
     return 50;
 };
 
@@ -45,11 +45,11 @@ bool get_PRESSURE() {
     return true;
 };
 
-int get_PRESSURE_MIN() {
+short int get_PRESSURE_MIN() {
     return 850;
 };
 
-int get_PRESSURE_MAX() {
+short int get_PRESSURE_MAX() {
     return 1080;
 };
 
@@ -58,11 +58,11 @@ uint16_t get_LOG_INTERVAL() {
     return 15;
 };
 
-int get_FILE_MAX_SIZE() {
+short int get_FILE_MAX_SIZE() {
     return 4096;
 };
 
-int get_TIMEOUT() {
+short int get_TIMEOUT() {
     return 30;
 };
